Moved options button to new file
* fixed header tint color for other user profile back button
