+++ home/components/OptionsButton.js
@@ -1,0 +1,50 @@
+/* @flow */
+import { connectActionSheet } from '@expo/react-native-action-sheet';
+import React from 'react';
+import { TouchableOpacity, StyleSheet, Alert, Platform } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+
+class OptionsButton extends React.Component {
+  render() {
+    return (
+      <TouchableOpacity style={styles.container} onPress={this._handlePress}>
+        <Ionicons
+          name={Platform.select({ ios: 'ios-more', default: 'md-more' })}
+          size={27}
+          color="#4E9BDE"
+        />
+      </TouchableOpacity>
+    );
+  }
+
+  _handlePress = () => {
+    let options = ['Report this user', 'Cancel'];
+    let cancelButtonIndex = 1;
+    this.props.showActionSheetWithOptions(
+      {
+        options,
+        cancelButtonIndex,
+      },
+      async buttonIndex => {
+        if (buttonIndex === 0) {
+          Alert.alert(
+            'Thank you for your report',
+            'We will investigate the case as soon as we can.'
+          );
+        }
+      }
+    );
+  };
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingRight: 15,
+  },
+});
+
+export default connectActionSheet(OptionsButton);
