+++ home/screens/ProfileScreen.js
@@ -1,51 +1,37 @@
 /* @flow */
-
+import { MaterialIcons } from '@expo/vector-icons';
 import React from 'react';
 import {
-  Alert,
+  findNodeHandle,
   NativeModules,
   Platform,
   StyleSheet,
   Text,
   TouchableOpacity,
   View,
-  findNodeHandle,
 } from 'react-native';
-import { Ionicons, MaterialIcons } from '@expo/vector-icons';
-import { connect } from 'react-redux';
-import { connectActionSheet } from '@expo/react-native-action-sheet';
 import { withNavigation } from 'react-navigation';
+import { connect } from 'react-redux';
 
-import SessionActions from '../redux/SessionActions';
-
+import OptionsButton from '../components/OptionsButton';
 import ProfileUnauthenticated from '../components/ProfileUnauthenticated';
 import MyProfileContainer from '../containers/MyProfileContainer';
 import OtherProfileContainer from '../containers/OtherProfileContainer';
-
+import SessionActions from '../redux/SessionActions';
 import getViewerUsernameAsync from '../utils/getViewerUsernameAsync';
-import onlyIfAuthenticated from '../utils/onlyIfAuthenticated';
 import isUserAuthenticated from '../utils/isUserAuthenticated';
+import onlyIfAuthenticated from '../utils/onlyIfAuthenticated';
 
 @connect((data, props) => ProfileScreen.getDataProps(data, props))
 export default class ProfileScreen extends React.Component {
   static navigationOptions = ({ navigation }) => ({
+    headerTintColor: '#4E9BDE',
     title: navigation.getParam('username', 'Profile'),
-    ...Platform.select({
-      ios: {
-        headerRight: navigation.getParam('username') ? (
-          <OptionsButtonIOS />
-        ) : (
-          <UserSettingsButtonIOS />
-        ),
-      },
-      android: {
-        headerRight: navigation.getParam('username') ? (
-          <OptionsButtonAndroid />
-        ) : (
-          <SignOutButtonAndroid />
-        ),
-      },
-    }),
+    headerRight: navigation.getParam('username') ? (
+      <OptionsButton />
+    ) : (
+      Platform.select({ ios: <UserSettingsButtonIOS />, default: <SignOutButtonAndroid /> })
+    ),
   });
 
   static getDataProps(data, props) {
@@ -133,44 +119,6 @@ class SignOutButtonAndroid extends React.Component {
   };
 }
 
-class OptionsButtonAndroid extends React.Component {
-  _anchor: View;
-
-  render() {
-    return (
-      <View style={{ flex: 1 }}>
-        <View
-          collapsable={false}
-          ref={view => {
-            this._anchor = view;
-          }}
-          style={{ position: 'absolute', top: 5, left: 0 }}
-        />
-        <TouchableOpacity style={styles.buttonContainer} onPress={this._handlePress}>
-          <MaterialIcons name="more-vert" size={27} color="#000" />
-        </TouchableOpacity>
-      </View>
-    );
-  }
-
-  _handlePress = () => {
-    let handle = findNodeHandle(this._anchor);
-    NativeModules.UIManager.showPopupMenu(
-      handle,
-      ['Report this user'],
-      () => {},
-      (action, selectedIndex) => {
-        if (selectedIndex === 0) {
-          Alert.alert(
-            'Thank you for your report',
-            'We will investigate the case as soon as we can.'
-          );
-        }
-      }
-    );
-  };
-}
-
 @onlyIfAuthenticated
 @withNavigation
 class UserSettingsButtonIOS extends React.Component {
@@ -200,36 +148,6 @@ class SignOutButtonIOS extends React.Component {
 
   _handlePress = () => {
     this.props.dispatch(SessionActions.signOut());
-  };
-}
-
-@connectActionSheet
-class OptionsButtonIOS extends React.Component {
-  render() {
-    return (
-      <TouchableOpacity style={styles.buttonContainer} onPress={this._handlePress}>
-        <Ionicons name="ios-more" size={27} color="#4E9BDE" />
-      </TouchableOpacity>
-    );
-  }
-
-  _handlePress = () => {
-    let options = ['Report this user', 'Cancel'];
-    let cancelButtonIndex = 1;
-    this.props.showActionSheetWithOptions(
-      {
-        options,
-        cancelButtonIndex,
-      },
-      async buttonIndex => {
-        if (buttonIndex === 0) {
-          Alert.alert(
-            'Thank you for your report',
-            'We will investigate the case as soon as we can.'
-          );
-        }
-      }
-    );
   };
 }
